# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Boiler Service Technician API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Soft-delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
  /appliances:
    get:
      operationId: listAppliances
      tags: [appliances]
      summary: List appliances
      parameters:
        - name: property_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Appliance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appliance"
    post:
      operationId: createAppliance
      tags: [appliances]
      summary: Create an appliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplianceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appliance"
  /appliances/{id}:
    get:
      operationId: getAppliance
      tags: [appliances]
      summary: Get appliance by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Appliance details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplianceDetail"
    patch:
      operationId: updateAppliance
      tags: [appliances]
      summary: Update appliance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplianceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appliance"
    delete:
      operationId: deleteAppliance
      tags: [appliances]
      summary: Soft-delete appliance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
  /breakdown-reports:
    post:
      operationId: createBreakdownReport
      tags: [breakdownReports]
      summary: Create a breakdown report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBreakdownReportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakdownReport"
  /breakdown-reports/{id}:
    get:
      operationId: getBreakdownReport
      tags: [breakdownReports]
      summary: Get breakdown report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Breakdown report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakdownReport"
    patch:
      operationId: updateBreakdownReport
      tags: [breakdownReports]
      summary: Update breakdown report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBreakdownReportBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakdownReport"
  /breakdown-reports/job/{jobId}:
    get:
      operationId: getBreakdownReportByJob
      tags: [breakdownReports]
      summary: Get breakdown report by job ID
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Breakdown report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakdownReport"
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, email, full_name, role, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, office_staff, technician]
        phone:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateProfileBody:
      type: object
      properties:
        full_name:
          type: string
        phone:
          type: ["string", "null"]
    Customer:
      type: object
      required: [id, first_name, last_name, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: ["string", "null"]
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        address_line1:
          type: ["string", "null"]
        address_line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CustomerDetail:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            properties:
              type: array
              items:
                $ref: "#/components/schemas/Property"
    CreateCustomerBody:
      type: object
      required: [first_name, last_name]
      properties:
        title:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
        mobile:
          type: string
        address_line1:
          type: string
        address_line2:
          type: string
        city:
          type: string
        county:
          type: string
        postcode:
          type: string
        notes:
          type: string
    UpdateCustomerBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        address_line1:
          type: ["string", "null"]
        address_line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
    Property:
      type: object
      required: [id, customer_id, address_line1, postcode, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        address_line1:
          type: string
        address_line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postcode:
          type: string
        property_type:
          type: ["string", "null"]
        occupancy_type:
          type: ["string", "null"]
        access_notes:
          type: ["string", "null"]
        boiler_location:
          type: ["string", "null"]
        flue_location:
          type: ["string", "null"]
        tank_location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        customer_name:
          type: ["string", "null"]
...
    Appliance:
      type: object
      required: [id, property_id, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serial_number:
          type: ["string", "null"]
        boiler_type:
          type: ["string", "null"]
        fuel_type:
          type: ["string", "null"]
        system_type:
          type: ["string", "null"]
        installation_date:
          type: ["string", "null"]
          format: date
        warranty_expiry:
          type: ["string", "null"]
          format: date
        burner_make:
          type: ["string", "null"]
        burner_model:
          type: ["string", "null"]
        nozzle_size:
          type: ["string", "null"]
        pump_pressure:
          type: ["string", "null"]
        controls:
          type: ["string", "null"]
        last_service_date:
          type: ["string", "null"]
          format: date
        next_service_due:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        property_address:
          type: ["string", "null"]
    ApplianceDetail:
      allOf:
        - $ref: "#/components/schemas/Appliance"
        - type: object
          properties:
            property:
              $ref: "#/components/schemas/Property"
            recent_jobs:
              type: array
              items:
                $ref: "#/components/schemas/JobListItem"
    CreateApplianceBody:
      type: object
      required: [property_id]
      properties:
        property_id:
          type: string
          format: uuid
        manufacturer:
          type: string
        model:
          type: string
        serial_number:
          type: string
        boiler_type:
          type: string
        fuel_type:
          type: string
        system_type:
          type: string
        installation_date:
          type: string
          format: date
        warranty_expiry:
          type: string
          format: date
        burner_make:
          type: string
        burner_model:
          type: string
        nozzle_size:
          type: string
        pump_pressure:
          type: string
        controls:
          type: string
        last_service_date:
          type: string
          format: date
        next_service_due:
          type: string
          format: date
        notes:
          type: string
    UpdateApplianceBody:
      type: object
      properties:
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serial_number:
          type: ["string", "null"]
        boiler_type:
          type: ["string", "null"]
        fuel_type:
          type: ["string", "null"]
        system_type:
          type: ["string", "null"]
        installation_date:
          type: ["string", "null"]
        warranty_expiry:
          type: ["string", "null"]
        burner_make:
          type: ["string", "null"]
        burner_model:
          type: ["string", "null"]
        nozzle_size:
          type: ["string", "null"]
        pump_pressure:
          type: ["string", "null"]
        controls:
          type: ["string", "null"]
        last_service_date:
          type: ["string", "null"]
        next_service_due:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
    Job:
      type: object
      required: [id, customer_id, property_id, job_type, status, priority, scheduled_date, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        appliance_id:
          type: ["string", "null"]
          format: uuid
        assigned_technician_id:
          type: ["string", "null"]
          format: uuid
        job_type:
          type: string
          enum: [service, breakdown, installation, inspection, follow_up]
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled, requires_follow_up]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: ["string", "null"]
        estimated_duration:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobListItem:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            customer_name:
              type: ["string", "null"]
            property_address:
              type: ["string", "null"]
            technician_name:
              type: ["string", "null"]
...
    BreakdownReport:
      type: object
      required: [id, job_id, technician_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        reported_fault:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        diagnostics_performed:
          type: ["string", "null"]
        findings:
          type: ["string", "null"]
        parts_required:
          type: ["string", "null"]
        temporary_fix:
          type: ["string", "null"]
        permanent_fix:
          type: ["string", "null"]
        appliance_safe:
          type: boolean
        return_visit_required:
          type: boolean
        return_visit_notes:
          type: ["string", "null"]
        additional_notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateBreakdownReportBody:
      type: object
      required: [job_id, technician_id]
      properties:
        job_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        reported_fault:
          type: string
        symptoms:
          type: string
        diagnostics_performed:
          type: string
        findings:
          type: string
        parts_required:
          type: string
        temporary_fix:
          type: string
        permanent_fix:
          type: string
        appliance_safe:
          type: boolean
        return_visit_required:
          type: boolean
        return_visit_notes:
          type: string
        additional_notes:
          type: string
    UpdateBreakdownReportBody:
      type: object
      properties:
        reported_fault:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        diagnostics_performed:
          type: ["string", "null"]
        findings:
          type: ["string", "null"]
        parts_required:
          type: ["string", "null"]
        temporary_fix:
          type: ["string", "null"]
        permanent_fix:
          type: ["string", "null"]
        appliance_safe:
          type: boolean
        return_visit_required:
          type: boolean
        return_visit_notes:
          type: ["string", "null"]
        additional_notes:
          type: ["string", "null"]
...