# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Boiler Service Technician API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: string;
  customer_id: string;
  address_line1: string;
  /** @nullable */
  address_line2?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  county?: string | null;
  postcode: string;
  /** @nullable */
  property_type?: string | null;
  /** @nullable */
  occupancy_type?: string | null;
  /** @nullable */
  access_notes?: string | null;
  /** @nullable */
  boiler_location?: string | null;
  /** @nullable */
  flue_location?: string | null;
  /** @nullable */
  tank_location?: string | null;
  /** @nullable */
  notes?: string | null;
  is_active: boolean;
  created_at: Date;
  updated_at: Date;
  /** @nullable */
  customer_name?: string | null;
}