# lib/api-zod/src/generated/types/updateApplianceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Boiler Service Technician API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateApplianceBody {
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serial_number?: string | null;
  /** @nullable */
  boiler_type?: string | null;
  /** @nullable */
  fuel_type?: string | null;
  /** @nullable */
  system_type?: string | null;
  /** @nullable */
  installation_date?: string | null;
  /** @nullable */
  warranty_expiry?: string | null;
  /** @nullable */
  burner_make?: string | null;
  /** @nullable */
  burner_model?: string | null;
  /** @nullable */
  nozzle_size?: string | null;
  /** @nullable */
  pump_pressure?: string | null;
  /** @nullable */
  controls?: string | null;
  /** @nullable */
  last_service_date?: string | null;
  /** @nullable */
  next_service_due?: string | null;
  /** @nullable */
  notes?: string | null;
  is_active?: boolean;
}